feat: [factory-argument-signatures-prd-mergeability] - Restore pkg/cli file-count compliance
