--- pkg/cli/main_test.go
@@ -1,26 +1,0 @@
-package cli
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"testing"
-)
-
-func TestMain(m *testing.M) {
-	homeDir, err := os.MkdirTemp("", "you-cli-test-home-*")
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "create cli test home: %v\n", err)
-		os.Exit(1)
-	}
-	defer func() {
-		_ = os.RemoveAll(homeDir)
-	}()
-
-	os.Setenv("HOME", homeDir)
-	os.Setenv("USERPROFILE", homeDir)
-	os.Setenv("HOMEDRIVE", filepath.VolumeName(homeDir))
-	os.Setenv("HOMEPATH", string(os.PathSeparator))
-
-	os.Exit(m.Run())
-}
