+++ pkg/cli/root_run_test.go
@@ -25,6 +25,24 @@ import (
 	"github.com/spf13/cobra"
 )
 
+func TestMain(m *testing.M) {
+	homeDir, err := os.MkdirTemp("", "you-cli-test-home-*")
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "create cli test home: %v\n", err)
+		os.Exit(1)
+	}
+	defer func() {
+		_ = os.RemoveAll(homeDir)
+	}()
+
+	os.Setenv("HOME", homeDir)
+	os.Setenv("USERPROFILE", homeDir)
+	os.Setenv("HOMEDRIVE", filepath.VolumeName(homeDir))
+	os.Setenv("HOMEPATH", string(os.PathSeparator))
+
+	os.Exit(m.Run())
+}
+
 func TestRunCommand_VerboseFlag(t *testing.T) {
 	root := NewRootCommand()
 	runCmd, _, err := root.Find([]string{"run"})
